address-book-rubrica: Declare it final
Use G_DECLARE_FINAL_TYPE for LibBalsaAddressBookRubrica and make object
and class opaque.

LibBalsaAddressBookRubrica made liberal use of the members of
LibBalsaAddressBookText, which are now private. The path can be changed
and has a getter and a setter, so ab_rubrica uses those. The other
members are essentially private to ab_rubrica, so the storage for those
is duplicated in the LibBalsaAddressBookRubrica object, a small price to
pay for the convenience of direct access.