mime-widget-text: Grab the focus when realized
The first time we click in a text part, the text widget grabs the
focus, and that leads the container that holds it to scroll to
its focus child, namely the text widget. That moves the text
under the pointer, so the click is received somewhere other than
was intended. Grabbing the focus on realization seems to avert
the scrolling.