mime-widget-text: The "draw" signal has gone away
We would like to create cite-bars at "draw" time, but that signal has
been removed. Connecting to "realize" or "map" seems to be too early in
the process, as the cite-bars are not drawn. Using an idle handler gets
them shown, so we'll use that fix. There may be a slight delay between
the message being rendered and the cite-bars being added to it, but for
now it seems to be a necessary price to pay.