Use GResource instead of installed files
Current Balsa uses installed ui/*.ui files to define the user interfaces
for the main window, the message window, the compose window, and the
view-source window. Those files can be replaced by GResource objects,
which are part of the binary file, so that accessing them requires no
additional file system access.

This commit:
• changes the C code to use GResources instead of installed files;
• changes various meson.build files to generate the necessary C files
  instead of installing XML ui files (not difficult);
• changes configure.ac and various Makefile.am files for the same
  purpose (far more difficult, for me!).