folder-conf: Use a cast instead of an address
We can cast FolderDialogData and SubfolderDialogData to
CommonDialogData, and we do, implicitly, in passing them to the destroy
handler, so for consistency we should do that elsewhere instead of
taking the address of the first member of each.

Taking the address does not depend on the common data being the first
member, but the implicit cast does.