revert renaming of buffer_to_surface_coords()
```
textview: revert renaming of buffer_to_surface_coords()

The renaming of this function doesn't make much since because the window
is the GtkTextWindowType, not GdkWindow specifically. So we can keep the
old name which is closer to the proper meaning and less code for consumers
to change when porting to 4.x.
```

See gtk commit 8373cc6c471f478dbf48fad7f5532e27f286c150